Only remove containers and volume if they exist

+++ test.sh
@@ -1,8 +1,16 @@
 #!/bin/bash
 
 # Delete the containers and the database
-docker rm remitly-api-1 remitly-database-1
-docker volume rm remitly_postgres-data
+if [ "$(docker ps -a -q -f name=remitly-api-1)" ]; then
+    docker rm remitly-api-1
+fi
+if [ "$(docker ps -a -q -f name=remitly-database-1)" ]; then
+    docker rm remitly-database-1
+fi
+
+if [ "$(docker volume ls -q -f name=remitly_postgres-data)" ]; then
+    docker volume rm remitly_postgres-data
+fi
 
 # Run the tests
 docker compose -f docker-compose.test.yml up --no-attach database
